Better NA checks in RTdist.
If there are TOTALLY missing output positions (i.e the simulation produced exactly zero recalls for a given memory search order), those positions will NOT be included in the outout. It is the caller's job to deal with the missing data, as those completely missing values should almost certainly not be left out of any likelihood calculations.