moved the subject and group level accuracy and RT analysis for the LB4L2 experiment out of the data-munging script and into specialized functions provided by the package.
The actuall summarizing function for accuracy and RT data is provided by an S3 method for summary, and the plotting of the group-level results is provided by an S3 method for autoplot.

Serveral helper functions are included to check the whether an object inherits from one of the classes in this package, and to coerce objects into one of the classes from this package.